requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==1.1.2
gunicorn==20.0.4
sqlalchemy==1.4.14
psycopg2==2.8.6
pyjwt==2.1.0
flask-sqlalchemy